pyproject.toml
[project]
name = "dldbt"
version = "0.1.0"
description = "Map git branches to DuckLake schemas. dbt runs respect the branch; merging reuses computed data."
readme = "README.md"
authors = [
    { name = "Jesse Ginsberg", email = "jginsberg3@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.2",
    "psycopg[binary]>=3.3.3",
    "typer>=0.13.0",
    "rich>=14.0.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
dldbt = "dldbt.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dldbt"]

[tool.hatch.build.targets.wheel.force-include]
"src/dldbt/dbt_ops/macros" = "dldbt/dbt_ops/macros"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
    "types-pyyaml>=6.0.12",
    "dbt-duckdb>=1.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires docker compose (postgres + minio) running",
]
